refactor(ui5-upload-collection-item): migrate to TypeScript (#6296)
Part of #4337
